refactor(frontend): Shared UI-Utilities für View-Komponenten extrahieren (#193)
renderStateBadge, renderLabel und renderAvatar in gemeinsames
Modul src/utils/render-helpers.ts extrahiert. issues-view und
pull-requests-view verwenden die gemeinsamen Funktionen.
